git-p4: python 3 syntax changes
Python 3 handles strings differently than Python 2.7.  Since Python 2
is reaching it's end of life, a series of changes are being submitted to
enable python 3.5 and following support. The current code fails basic
tests under python 3.5.

There are a number of translations suggested by modernize/futureize that
should be taken to fix numerous non-string specific issues.

Change references to the X.next() iterator to the function next(X) which
is compatible with both Python2 and Python3.

Change references to X.keys() to list(X.keys()) to return a list that
can be iterated in both Python2 and Python3.

Add the literal text (object) to the end of class definitions to be
consistent with Python3 class definition.

Change integer divison to use "//" instead of "/"  Under Both Python 2
and Python 3 // will return a floor()ed result which matches existing
functionality.

Change the format string for displaying decimal values from %d to %4.1f%
when displaying a progress.  This avoids displaying long repeating
decimals in user displayed text.

Signed-off-by: Ben Keene <seraphire@gmail.com>