git-p4: p4CmdList - support Unicode encoding
The p4CmdList is a commonly used function in the git-p4 code. It is used
to execute a command in P4 and return the results of the call in a list.

Change this code to take a new optional parameter, encode_data that will
optionally convert the data AS_STRING() that isto be returned by the
function.

Change the code so that the key will always be encoded as_string()

Data that is passed for standard input (stdin) should be as_bytes() to
ensure unicode text that is supplied will be written out as bytes.

Additionally, change literal text prior to conversion to be literal
bytes.

Signed-off-by: Ben Keene <seraphire@gmail.com>